--- feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/Admin.java
@@ -1,5 +1,0 @@
-package com.equipo5.feelflowapp.domain;
-
-public class Admin extends User{
-
-}
