+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/Team.java
@@ -1,12 +1,37 @@
 package com.equipo5.feelflowapp.domain;
 
+import com.equipo5.feelflowapp.domain.users.RegularUser;
+import com.equipo5.feelflowapp.domain.users.TeamLeader;
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class Team {
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(name="UUID",strategy="org.hibernate.id.UUIDGenerator")
+    @JdbcTypeCode(SqlTypes.CHAR)
+    @Column(length = 36,columnDefinition = "varchar(36)",updatable = false,nullable = false)
     private UUID uuid;
+
     private String name;
     private String descriptionProject;
+    private EnterPrise enterPrise;
+    private TeamLeader teamLeader;
+
+    @Builder.Default
     private List<RegularUser> regularUsers = new ArrayList<>();
 }
