+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/enumerations/teamRoles/TeamRoles.java
@@ -1,5 +1,5 @@
 package com.equipo5.feelflowapp.domain.enumerations.teamRoles;
 
 public enum TeamRoles {
-    ADMINISTRADOR,TEAM_LEADER,USER_REGULAR
+    ADMIN,TEAM_LEADER,USER_REGULAR
 }
