+++ feel-flow-package/src/main/java/com/equipo5/feelflowapp/domain/users/User.java
@@ -1,16 +1,13 @@
-package com.equipo5.feelflowapp.domain;
+package com.equipo5.feelflowapp.domain.users;
 
-import com.equipo5.feelflowapp.domain.users.Admin;
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
+
+import com.equipo5.feelflowapp.domain.enumerations.teamRoles.TeamRoles;
+import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 @Getter
@@ -18,7 +15,8 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class EnterPrise {
+@Table(name = "user_model")
+public class User {
     @Id
     @GeneratedValue(generator = "UUID")
     @GenericGenerator(name="UUID",strategy="org.hibernate.id.UUIDGenerator")
@@ -27,8 +25,9 @@ public class EnterPrise {
     private UUID uuid;
 
     private String name;
-    private Admin admin;
+    private String email;
+    private String password;
 
-    @Builder.Default
-    private List<Team> team = new ArrayList<>();
+    @Enumerated(EnumType.STRING)
+    private TeamRoles role;
 }
